Attempt at finding libraries used by a lineage
Able to find the libraries, but we really need to know what lines of
code to add to load the libraries.  Probably the best fix is to have
rdtLite add a proc id in the library nodes indicating where the library
call occurs (if at all).  It might be loaded using :: syntax, in which
case library is not needed, or it might be loaded inside a function or
some other library code that we do not see, and would hopefully (!) be
included in the lineage already, but is that guaranteed??